P1: add global Attention Inbox to the Session rail (#399)
Wire the capability-driven global Attention Inbox into the Session-first rail while preserving native Session discovery, transcript and writer boundaries.
