+++ web/src/session-first-regression.test.mjs
@@ -11,7 +11,7 @@ const adapter = read('./native-session-v3-adapter.ts')
 const conversationController = read('./conversation-controller.ts')
 const modelRecovery = read('./native-session-model.ts')
 const observer = read('./components/native-session-observer.tsx')
-const home = read('./components/native-session-home.tsx')
+const home = read('./components/native-session-home-base.tsx')
 const actions = read('./components/native-session-actions.tsx')
 const rename = read('./components/native-session-rename.tsx')
 const workThread = read('./components/work-thread-conversation.tsx')
