+++ web/src/ui-regression.test.mjs
@@ -5,6 +5,8 @@ const read = (name) => readFileSync(new URL(name, import.meta.url), 'utf8')
 const main = read('./main.tsx')
 const shell = read('./components/standalone-universal-workspace.tsx')
 const home = read('./components/native-session-home.tsx')
+const homeBase = read('./components/native-session-home-base.tsx')
+const attentionInbox = read('./components/native-session-home-attention.tsx')
 const observer = read('./components/native-session-observer.tsx')
 const chat = read('./components/work-thread-conversation.tsx')
 const shared = read('./components/taskdesk-conversation.tsx')
@@ -18,9 +20,15 @@ assert.equal(api.includes('const text = await response.text()'), false, 'error h
 assert.match(main, /<StandaloneUniversalWorkspace/, 'the product must boot directly into Session-first')
 assert.match(shell, /<NativeSessionHome/, 'the Session rail must be native discovery')
 assert.match(shell, /<NativeSessionObserver/, 'the detail pane must open the native Session surface')
-assert.match(home, /hr-native-machine-group/, 'Session navigation must preserve machine grouping')
-assert.match(home, /hr-native-project-group/, 'Session navigation must preserve Project grouping')
-assert.match(home, /hr-native-session-row/, 'Session navigation must render native Session rows')
+assert.match(home, /native-session-home-attention/, 'the public Session rail entrypoint must compose the global Attention Inbox')
+assert.match(homeBase, /hr-native-machine-group/, 'Session navigation must preserve machine grouping')
+assert.match(homeBase, /hr-native-project-group/, 'Session navigation must preserve Project grouping')
+assert.match(homeBase, /hr-native-session-row/, 'Session navigation must render native Session rows')
+assert.match(attentionInbox, /loadNativeSessionAttentionIndex/, 'the Inbox must use the capability-driven attention read model')
+assert.match(attentionInbox, /startNativeSessionAttentionLiveRefresh/, 'the Inbox must use its dedicated capability-scoped live refresh')
+assert.match(attentionInbox, /baseAttentionCount \+ inbox\.length/, 'mobile navigation must carry explicit global attention as well as legacy rail attention')
+assert.match(attentionInbox, /onAttentionCountChange=\{setBaseAttentionCount\}/, 'the wrapper must intercept the base count before forwarding the combined badge count')
+assert.doesNotMatch(attentionInbox, /startTaskDeskSessionLiveRefresh|loadMessagePage|continueConversation|stopConversation/, 'global attention must not reuse transcript or Session writer paths')
 
 assert.match(observer, /<WorkThreadConversation/, 'native Session detail must reuse the mature v3 controller')
 assert.match(chat, /<TaskDeskConversation/, 'the mature controller must own the shared transcript/composer')
