+++ web/src/v3-offline-visibility.test.mjs
@@ -4,7 +4,8 @@ import test from "node:test"
 
 const read = (name) => readFileSync(new URL(name, import.meta.url), "utf8").replace(/\r\n/g, "\n")
 const workspace = read("./components/standalone-universal-workspace.tsx")
-const home = read("./components/native-session-home.tsx")
+const home = read("./components/native-session-home-base.tsx")
+const attentionHome = read("./components/native-session-home-attention.tsx")
 const attention = read("./components/work-thread-attention.tsx")
 const workbench = read("./session-first-workbench.css")
 const manifest = JSON.parse(read("../public/manifest.webmanifest"))
@@ -53,6 +54,7 @@ test("the attention surface is announced and its options report their state", ()
   assert.match(attention, /aria-live="polite"/)
   assert.match(attention, /aria-pressed=\{selected\.includes\(option\.label\)\}/)
   assert.match(attention, /aria-label=\{`Custom answer for/)
+  assert.match(attentionHome, /aria-label="Attention Inbox"/, "global pending attention must be announced separately from the selected Session detail")
 })
 
 test("the installed app describes the multi-harness product", () => {
@@ -65,6 +67,7 @@ test("native Session search and attention filtering stay reachable on phone", ()
   assert.match(home, /aria-label=\{t\("sf\.searchSessionsLabel"\)\}/)
   assert.match(home, /filter === "attention"/)
   assert.match(home, /t\("sf\.filterAttention"\)/)
+  assert.match(attentionHome, /Authorization required/, "global authorization requests must remain visible on the Session rail")
   assert.match(workspace, /hr-mobile-nav-badge/)
   assert.match(workspace, /onAttentionCountChange=\{setAttentionCount\}/)
 })
