+++ web/src/v3-ux-polish-regression.test.mjs
@@ -10,7 +10,8 @@ const exists = (name) => fs.existsSync(path.join(here, name))
 const profiles = read("serverProfiles.ts")
 const main = read("main.tsx")
 const standalone = read("components/standalone-universal-workspace.tsx")
-const home = read("components/native-session-home.tsx")
+const home = read("components/native-session-home-base.tsx")
+const attentionHome = read("components/native-session-home-attention.tsx")
 const observer = read("components/native-session-observer.tsx")
 const chat = read("components/work-thread-conversation.tsx")
 const conversationSurface = read("components/taskdesk-conversation.tsx")
@@ -47,12 +48,15 @@ assert.doesNotMatch(standalone, /New conversation/i)
 assert.doesNotMatch(standalone, />Conversations</)
 assert.doesNotMatch(standalone, /TaskDeskWorkspace/)
 
-// The navigation hierarchy remains Machine -> Project -> native Session.
+// The navigation hierarchy remains Machine -> Project -> native Session, with global Attention
+// composed above it rather than replacing the mature rail.
 assert.match(home, /hr-native-machine-group/)
 assert.match(home, /hr-native-project-group/)
 assert.match(home, /hr-native-session-row/)
 assert.match(home, /sessionTreeRows/)
 assert.match(home, /createNativeSessionTarget/)
+assert.match(attentionHome, /<NativeSessionHomeBase/)
+assert.match(attentionHome, /hr-native-attention-inbox/)
 assert.match(observer, /<WorkThreadConversation/)
 assert.match(chat, /<TaskDeskConversation/)
 assert.match(sessionWorkbench, /hr-native-workspace/)
